+++ Billbee.Api.Client/Model/UpdateStockCode.cs
@@ -3,7 +3,7 @@
     public class UpdateStockCode
     {
         public string Sku { get; set; }
-        public int? StockId { get; set; }
+        public long? StockId { get; set; }
         public string StockCode { get; set; }
     }
 }
